+++ examples/celldevs_sir/simulation/celldevs_sir_sequential_root_coordinator.hpp
@@ -27,7 +27,7 @@
 #include <thread>
 #include <fstream>
 
-void sequential_simulation(size_t n_subcomponents, CellDEVSSirAtomic* subcomponents, size_t* n_couplings, int** couplings , size_t sim_time) {
+void sequential_simulation(size_t n_subcomponents, CellDEVSSirAtomic* subcomponents, size_t* n_couplings, size_t** couplings , size_t sim_time) {
 	double next_time = 0, last_time = 0;
 
 	while(next_time < sim_time) {
